Remove unused Operations library
Apart from the Batch extension method, this is unused by Epro. Anything
else still using it should be modified to not use it ASAP.

refs EP-45015
