--- LargeCollectionProfiler/LargeCollectionProfiler.csproj
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Remove="*.ncrunchproject" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\LargeCollections.FileSystem\LargeCollections.FileSystem.csproj" />
-    <ProjectReference Include="..\LargeCollections.Operations\LargeCollections.Operations.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="NConsoler">
-      <Version>2.0.0</Version>
-    </PackageReference>
-  </ItemGroup>
-</Project>
