+++ LargeCollections.Core/Operations/Extensions.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+
+namespace LargeCollections.Core.Operations
+{
+    public static class Extensions
+    {
+        public static IEnumerator<IEnumerable<T>> Batch<T>(this IEnumerator<T> enumerator, int batchSize)
+        {
+            return new BatchedSinglePassCollection<T>(enumerator, batchSize).InheritsCount(enumerator);
+        }
+
+        public static IEnumerator<IEnumerable<T>> Batch<T>(this ILargeCollection<T> collection, int batchSize)
+        {
+            return new BatchedSinglePassCollection<T>(collection.GetEnumerator(), batchSize).InheritsCount(collection);
+        }
+    }
+}
