--- LargeCollections.Operations.Tests/LargeCollections.Operations.Tests.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net462</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Moq" Version="4.13.1" />
-    <PackageReference Include="NUnit" Version="3.12.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\LargeCollections.FileSystem\LargeCollections.FileSystem.csproj" />
-    <ProjectReference Include="..\LargeCollections.Operations\LargeCollections.Operations.csproj" />
-    <ProjectReference Include="..\LargeCollections.Tests\LargeCollections.Tests.csproj" />
-  </ItemGroup>
-
-</Project>
