--- LargeCollections.Operations/ISortedMerge.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace LargeCollections.Operations
-{
-    public interface ISortedMerge<T>
-    {
-        IEnumerator<T> Merge(IEnumerable<IEnumerator<T>> enumerators);
-    }
-}
