--- LargeCollections.Operations/LargeCollections.Operations.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net462</TargetFrameworks>
-    <VersionPrefix>1.0.0</VersionPrefix>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\LargeCollections.Core\LargeCollections.Core.csproj" />
-  </ItemGroup>
-
-</Project>
