--- LargeCollections.Operations/SetUnionSortPreservingMerge.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace LargeCollections.Operations
-{
-    public class SetUnionSortPreservingMerge<T> : SortPreservingMergeBase<T>
-    {
-        protected override IEnumerator<T> MergeEnumerators(SortedEnumeratorList<T> sortedEnumerators)
-        {
-            if (sortedEnumerators.AdvanceAll())
-            {
-                do
-                {
-                    var first = sortedEnumerators.First();
-                    yield return first.Current;
-                    sortedEnumerators.Advance(first);
-                } while (sortedEnumerators.Any());
-            }
-        }
-    }
-}
