+++ LargeCollections.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.Tests", "LargeCollections.Tests\LargeCollections.Tests.csproj", "{476F7C46-D589-408D-907E-4971B2722BBB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollectionProfiler", "LargeCollectionProfiler\LargeCollectionProfiler.csproj", "{4352A39E-54EC-4E7C-B05E-D8C042EF688D}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SolutionPackages", "SolutionPackages\SolutionPackages.csproj", "{71191707-58E7-4BC8-B035-259D7DCEC50F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bluewire.ReferenceCounting", "Bluewire.ReferenceCounting\Bluewire.ReferenceCounting.csproj", "{2CF95D96-7B41-4E0E-9FE3-1D6085B08BA4}"
@@ -15,15 +13,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.SqlServer"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.FileSystem", "LargeCollections.FileSystem\LargeCollections.FileSystem.csproj", "{07A2BBE5-0CB9-4EA6-8386-4DAE9397D3BA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.Operations", "LargeCollections.Operations\LargeCollections.Operations.csproj", "{90A84227-E6A0-4141-A9F8-B0FCD08029C9}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.Core", "LargeCollections.Core\LargeCollections.Core.csproj", "{5FCCB031-DD19-40AA-9EDF-EEBFA582B8D2}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.SqlServer.Tests", "LargeCollections.SqlServer.Tests\LargeCollections.SqlServer.Tests.csproj", "{DFBF940C-6496-4AC0-84D4-86CA48139240}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.FileSystem.Tests", "LargeCollections.FileSystem.Tests\LargeCollections.FileSystem.Tests.csproj", "{DFCA93D5-C502-4C6A-8985-14DC482648EB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeCollections.Operations.Tests", "LargeCollections.Operations.Tests\LargeCollections.Operations.Tests.csproj", "{B22058CA-0F64-4B3F-AF81-986C7091987C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bluewire.ReferenceCounting.Tests", "Bluewire.ReferenceCounting.Tests\Bluewire.ReferenceCounting.Tests.csproj", "{A084E680-3A2F-4004-88BB-6AE3C70B4D4E}"
 EndProject
@@ -47,18 +41,6 @@ Global
 		{476F7C46-D589-408D-907E-4971B2722BBB}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{476F7C46-D589-408D-907E-4971B2722BBB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{476F7C46-D589-408D-907E-4971B2722BBB}.Release|x86.ActiveCfg = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Debug|x86.Build.0 = Debug|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|x86.ActiveCfg = Release|Any CPU
-		{4352A39E-54EC-4E7C-B05E-D8C042EF688D}.Release|x86.Build.0 = Release|Any CPU
 		{71191707-58E7-4BC8-B035-259D7DCEC50F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{71191707-58E7-4BC8-B035-259D7DCEC50F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{71191707-58E7-4BC8-B035-259D7DCEC50F}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -105,18 +87,6 @@ Global
 		{07A2BBE5-0CB9-4EA6-8386-4DAE9397D3BA}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{07A2BBE5-0CB9-4EA6-8386-4DAE9397D3BA}.Release|x86.ActiveCfg = Release|Any CPU
 		{07A2BBE5-0CB9-4EA6-8386-4DAE9397D3BA}.Release|x86.Build.0 = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Debug|x86.Build.0 = Debug|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|x86.ActiveCfg = Release|Any CPU
-		{90A84227-E6A0-4141-A9F8-B0FCD08029C9}.Release|x86.Build.0 = Release|Any CPU
 		{5FCCB031-DD19-40AA-9EDF-EEBFA582B8D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5FCCB031-DD19-40AA-9EDF-EEBFA582B8D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FCCB031-DD19-40AA-9EDF-EEBFA582B8D2}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -153,18 +123,6 @@ Global
 		{DFCA93D5-C502-4C6A-8985-14DC482648EB}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{DFCA93D5-C502-4C6A-8985-14DC482648EB}.Release|x86.ActiveCfg = Release|Any CPU
 		{DFCA93D5-C502-4C6A-8985-14DC482648EB}.Release|x86.Build.0 = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Debug|x86.Build.0 = Debug|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|x86.ActiveCfg = Release|Any CPU
-		{B22058CA-0F64-4B3F-AF81-986C7091987C}.Release|x86.Build.0 = Release|Any CPU
 		{A084E680-3A2F-4004-88BB-6AE3C70B4D4E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A084E680-3A2F-4004-88BB-6AE3C70B4D4E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A084E680-3A2F-4004-88BB-6AE3C70B4D4E}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
